requirements.txt
requests
pywebview
beautifulsoup4
google-generativeai